+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 * Resources build phase generation https://github.com/xcbuddy/xcbuddy/pull/38 by @pepibumur.
 * Headers build phase generation https://github.com/xcbuddy/xcbuddy/pull/41 by @pepibumur.
 * Shell script build phase generation https://github.com/xcbuddy/xcbuddy/pull/43 by @pepibumur.
+* Copy files build phase generation https://github.com/xcbuddy/xcbuddy/pull/44 by @pepibumur.
 
 ### Changed